requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
oauthlib==0.6.3
requests==2.3.0
requests-oauthlib==0.4.1
wsgiref==0.1.2
